Dev swiftlint corelib

+++ LGKKCoreLib/LGKKCoreLib/Modules/SPExtensions/Realm+SPExtension.swift
@@ -10,7 +10,7 @@ import UIKit
 import Realm
 import RealmSwift
 
-//Write helper
+// Write helper
 public extension Realm {
     
     public func safeAdd(object: Object) {
@@ -25,8 +25,8 @@ public extension Realm {
                 try realm.commitWrite()
             } else {
                 Realm.removeOldRealmDataFile()
-                //Re-try again
-                let realm = try! Realm()
+                // Re-try again
+                let realm = try Realm()
                 realm.beginWrite()
                 try block(realm)
                 try realm.commitWrite()
@@ -37,19 +37,22 @@ public extension Realm {
     }
     
     public class func removeOldRealmDataFile() {
-        let realmURL = Realm.Configuration.defaultConfiguration.fileURL!
-        let realmURLs = [
-            realmURL,
-            realmURL.appendingPathExtension("lock"),
-            realmURL.appendingPathExtension("note"),
-            realmURL.appendingPathExtension("management")
-        ]
-        for URL in realmURLs {
-            do {
-                try FileManager.default.removeItem(at: URL)
-            } catch {
-                print("Still failed when try to remove old realm data files.")
+        if let realmURL = Realm.Configuration.defaultConfiguration.fileURL {
+            let realmURLs = [
+                realmURL,
+                realmURL.appendingPathExtension("lock"),
+                realmURL.appendingPathExtension("note"),
+                realmURL.appendingPathExtension("management")
+            ]
+            for URL in realmURLs {
+                do {
+                    try FileManager.default.removeItem(at: URL)
+                } catch {
+                    print("Still failed when try to remove old realm data files.")
+                }
             }
+        } else {
+            print("Still failed when find realm url.")
         }
     }
     
@@ -59,8 +62,8 @@ public extension Realm {
                 try block(realm)
             } else {
                 Realm.removeOldRealmDataFile()
-                //Re-try again
-                let realm = try! Realm()
+                // Re-try again
+                let realm = try Realm()
                 try block(realm)
             }
         } catch {
@@ -68,4 +71,3 @@ public extension Realm {
         }
     }
 }
-
